fix: responsive styles added
