+++ src/components/Teaser.jsx
@@ -41,7 +41,7 @@ const Teaser = () => {
         <span className='pill'>Magazine</span>
         <span className='pill'>Surfing</span>
       </div>
-      <Link to={`/articles/article:aloha-spirits-in-northern-norway`}>
+      <Link to={`/articles/article:aloha-spirits-in-northern-norway${window.location.search}`}>
         <button>Read more</button>
       </Link>
     </article>
