+++ src/components/Teaser.scss
@@ -42,3 +42,16 @@ it.
     opacity: 0.6;
   }
 }
+
+@media only screen and (max-width: $tablet-breakpoint) {
+  .Teaser {
+    padding: $tablet-padding;
+  }
+}
+
+@media only screen and (max-width: $mobile-breakpoint) {
+  .Teaser {
+    display: flex;
+    flex-direction: column-reverse;
+  }
+}
