+++ src/styles/_variables.scss
@@ -65,3 +65,5 @@ $gutter-padding: 12px;
 // Breakpoints
 $mobile-breakpoint: 768px;
 $tablet-breakpoint: 1024px;
+
+$tablet-padding: 40px;
